Data Preparation
Source:
The dataset is publicly available on the kaggle website, and was collected from hospitals from an ongoing study on the residents of the town of Framingham. The intention is to replicate the analysis from this data fo high risk patients in regions across the sub-Sahara.

https://www.kaggle.com/amanajmera1/framingham-heart-study-dataset/data
The classification goal is to predict whether a patient has a 10-year risk of future coronary heart disease (CHD). The dataset provides the patients' information. It includes over 4,000 records and 15 attributes.